fix(xtask): split update-status structural modules (#0000)
